refactor: [WC-52] ScrollGuide컴포넌트 스타일 병합
